LayoutTests/imported/w3c/web-platform-tests/webxr/hit-test/ar_hittest_subscription_refSpaces.https-expected.txt


FAIL Ensures subscription to hit test works with viewer space - straight ahead - plane - webgl assert_equals: Results length should match expected results length expected 1 but got 0
FAIL Ensures subscription to hit test works with viewer space - straight ahead - plane - webgl2 assert_equals: Results length should match expected results length expected 1 but got 0
PASS Ensures subscription to hit test works with viewer space - straight up - no results - webgl
PASS Ensures subscription to hit test works with viewer space - straight up - no results - webgl2
FAIL Ensures subscription to hit test works with local space - webgl assert_equals: Results length should match expected results length expected 1 but got 0
FAIL Ensures subscription to hit test works with local space - webgl2 assert_equals: Results length should match expected results length expected 1 but got 0
FAIL Ensures subscription to hit test works with local-floor space - webgl assert_equals: Results length should match expected results length expected 1 but got 0
FAIL Ensures subscription to hit test works with local-floor space - webgl2 assert_equals: Results length should match expected results length expected 1 but got 0
